test(cli): use the shared stripAnsi helper across test files
Consolidates 16 near-duplicate local stripAnsi definitions onto a single
helper. Moved it into its own tests/helpers/ansi.ts module (re-exported
from tests/helpers/cli.ts for existing e2e callers) so unit/integration
tests don't drag in cli.ts's subprocess-spawning e2e machinery. The
shared implementation strips any CSI escape sequence, matching
agent-output.e2e.test.ts's stricter cursor-code-aware version rather
than the color-only \x1b[...m regex most copies used.
