+++ apps/cli/src/legacy/cli/agent-output.e2e.test.ts
@@ -1,26 +1,5 @@
 import { describe, expect, test } from "vitest";
-import { runSupabase } from "../../../tests/helpers/cli.ts";
-
-function stripAnsi(output: string): string {
-  let stripped = "";
-  for (let i = 0; i < output.length; i++) {
-    const charCode = output.charCodeAt(i);
-    if (charCode !== 0x1b || output[i + 1] !== "[") {
-      stripped += output[i];
-      continue;
-    }
-
-    i += 2;
-    while (i < output.length) {
-      const code = output.charCodeAt(i);
-      if (code >= 0x40 && code <= 0x7e) {
-        break;
-      }
-      i++;
-    }
-  }
-  return stripped;
-}
+import { runSupabase, stripAnsi } from "../../../tests/helpers/cli.ts";
 
 function parseJsonLines(output: string): Array<unknown> {
   return stripAnsi(output)
