+++ apps/cli/src/legacy/commands/migration/list/list.integration.test.ts
@@ -4,6 +4,7 @@ import { BunServices } from "@effect/platform-bun";
 import { describe, expect, it } from "@effect/vitest";
 import { Cause, Effect, Exit, Layer, Option } from "effect";
 
+import { stripAnsi } from "../../../../../tests/helpers/ansi.ts";
 import {
   LEGACY_VALID_REF,
   mockLegacyCliConfig,
@@ -107,9 +108,6 @@ function setup(workdir: string, opts: SetupOpts = {}) {
     resolverCalls,
   };
 }
-
-// eslint-disable-next-line no-control-regex
-const stripAnsi = (text: string) => text.replace(/\x1b\[[0-9;]*m/gu, "");
 
 const flags = (over: Partial<LegacyMigrationListFlags> = {}): LegacyMigrationListFlags => ({
   dbUrl: over.dbUrl ?? Option.none(),
